refactor(frontend): extract useEpgPreview to its own hook file
